HR_Portal_Backend/requirements.txt
Werkzeug==3.0.1
jmespath==1.0.1
SQLAlchemy==2.0.23
requests==2.31.0
aniso8601==9.0.1
numpy==1.26.0
python-dotenv==1.0.0
boto3==1.28.84
protobuf==4.21.12
python-dateutil==2.8.2
rsa==4.9
pytz==2023.3.post1
pandas==2.1.3
blinker==1.7.0
six==1.16.0
s3transfer==0.7.0
ecdsa==0.18.0
Flask==3.0.0
itsdangerous==2.1.2
colorama==0.4.6
Flask-Cors==4.0.0
python-jose==3.3.0
typing_extensions==4.8.0
MarkupSafe==2.1.3
urllib3==2.0.7
Jinja2==3.1.2
Flask-RESTful==0.3.10
botocore==1.31.84
click==8.1.7
greenlet==3.0.1
pyasn1==0.5.0
mysql-connector-python==8.2.0
scipy==1.11.3
matplotlib==3.8.0